Adds a method to run native autoit commands; Put Control in a module
