--- bin/autoit
@@ -1,3 +1,0 @@
-#!/usr/bin/env ruby
-
-puts "ola"
